+++ src/algorithm/mod.rs
@@ -1,1 +1,2 @@
 pub mod delaunay_alg;
+pub mod delaunay_struct;
