+++ src/bin/soft_todelaunay.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use clap::Parser;
 use std::time::Instant;
 
-use compact_skel_3d::boundary3d::io;
+use compact_skel_3d::mesh3d::io;
 use compact_skel_3d::algorithm::delaunay_alg;
 
 #[derive(Parser)]
@@ -20,25 +20,25 @@ fn main() -> Result<()> {
     let obj_out_path_str = args.obj_out_path.to_str().unwrap_or("");
     
     println!("Loading topo_mesh");
-    let mut topomesh = io::load_obj(obj_path_str)?;
+    let mut mesh = io::load_obj(obj_path_str)?;
     
-    println!("Checking topo_mesh");
-    topomesh.check_topo_mesh()?;
+    println!("Checking mesh");
+    mesh.check_mesh()?;
     
     println!("Mesh to delaunay");
     let now = Instant::now();
-    delaunay_alg::to_delaunay(&mut topomesh, Some(std::f32::consts::PI * 20.0 / 180.0))?;
+    delaunay_alg::to_delaunay(&mut mesh, Some(std::f32::consts::PI * 20.0 / 180.0))?;
     let duration = now.elapsed();
     let sec = duration.as_secs();
     let min = sec / 60;
     let sec = sec - min * 60;
     println!("Delaunay mesh computed in {}m{}s", min, sec);
     
-    println!("Checking topo_mesh");
-    topomesh.check_topo_mesh()?;
+    println!("Checking mesh");
+    mesh.check_mesh()?;
 
     println!("Save mesh");
-    io::save_obj(obj_out_path_str, &topomesh)?;
+    io::save_obj(obj_out_path_str, &mesh)?;
     
     Ok(())
 }
