--- src/boundary3d/mod.rs
@@ -1,4 +1,0 @@
-pub mod topo_mesh;
-pub mod io;
-pub mod delaunay_struct;
-pub use topo_mesh::TopoMesh;
