+++ src/mesh3d/mod.rs
@@ -1,0 +1,4 @@
+pub mod mesh3d;
+pub mod io;
+pub mod mesh_operations;
+pub use mesh3d::Mesh3D;
